+++ src/main/scala/xitrum/util/PathSanitizer.scala
@@ -9,10 +9,7 @@ object PathSanitizer {
     val path2 = path.replace('\\', File.separatorChar).replace('/', File.separatorChar)
 
     // Simplistic dumb security check
-    if (path2.contains(File.separator + ".") ||
-        path2.contains("." + File.separator) ||
-        path2.startsWith(".")                ||
-        path2.endsWith(".")) {
+    if (path2.contains("..")) {
       None
     } else {
       Some(path2)